+++ languages/java8/template/src/main/java/algorithmia/algorithm/AlgorithmBasic.java
@@ -1,14 +1,14 @@
 package algorithmia.algorithm;
+
 import com.algorithmia.algorithmHandler.*;
 import com.algorithmia.*;
 
-class AlgorithmBasic
-{
-    String Apply(String input) throws Exception{
+class AlgorithmBasic {
+    String Apply(String input) throws Exception {
         return "hello ".concat(input);
     }
 
-    public static void main(String[] args) throws Exception{
+    public static void main(String[] args) throws Exception {
         AlgorithmBasic defs = new AlgorithmBasic();
         AlgorithmHandler algo = new AlgorithmHandler<>(defs, defs::Apply);
         algo.run();
